ryle_pipeline/requirements.txt
# Ryle Pipeline - Dependencies
google-genai>=1.0.0
supabase>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
typer[all]>=0.12.0
rich>=13.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
